+++ libs/state/spec/rx-state.docs.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { rxState } from './rx-state';
+import { rxState } from '../src/lib/rx-state';
 import { Component, Injectable } from '@angular/core';
 import { Subject, map, merge } from 'rxjs';
 import { AsyncPipe } from '@angular/common';