+++ libs/state/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [16.3.0](https://github.com/rx-angular/rx-angular/compare/state@16.2.0...state@16.3.0) (2023-11-01)
+
+
+### Features
+
+* expose select and get in rxState setup fn ([d04e465](https://github.com/rx-angular/rx-angular/commit/d04e4657a7e0b060895b5298cdc7a7a138474e98))
+
+
+
+# [16.2.0](https://github.com/rx-angular/rx-angular/compare/state@16.1.1...state@16.2.0) (2023-11-01)
+
+
+### Bug Fixes
+
+* build errors ([e8ff6d9](https://github.com/rx-angular/rx-angular/commit/e8ff6d913a6a1a01e9bd4f9356c2649b148664b9))
+* remove eslint comment ([c166c59](https://github.com/rx-angular/rx-angular/commit/c166c59d43fcdc0d33f2790103fc9becc2d336a9))
+
+
+### Features
+
+* added computedFrom impl ([7e80dc9](https://github.com/rx-angular/rx-angular/commit/7e80dc980bd7e9ed192ef7c1f3472f55d261a362))
+* added computedFrom impl ([8e2f578](https://github.com/rx-angular/rx-angular/commit/8e2f5784cfb6b50cd440d37fc1fe30151876541d))
+* **state:** introduce signal APIs for RxState ([d42888d](https://github.com/rx-angular/rx-angular/commit/d42888d2ffeac0808af7538f240698ac5f3bf672))
+* **state:** introduce signal APIs for RxState ([ac4b5dc](https://github.com/rx-angular/rx-angular/commit/ac4b5dc21a8e4a21fd63af63abb455ca7be9daf6))
+
+
+
 ## [16.1.1](https://github.com/rx-angular/rx-angular/compare/state@16.1.0...state@16.1.1) (2023-10-20)